Rename wso2.controlplane -> wso2.icp
