+++ .github/workflows/publish-release.yml
@@ -12,5 +12,5 @@ jobs:
     uses: ballerina-platform/ballerina-library/.github/workflows/release-package-connector-template.yml@main
     secrets: inherit
     with:
-      package-name: wso2.controlplane
+      package-name: wso2.icp
       package-org: ballerinax
