+++ ballerina-tests/artifacts-tests/tests/Config.toml
@@ -1,12 +1,12 @@
 testURL = "https://localhost:9165"
 testPort = 9090
 
-[ballerinax.wso2.controlplane]
+[ballerinax.wso2.icp]
 keyStorePath = "../resources/certsandkeys/ballerinaKeystore.p12"
 trustStorePath = "../resources/certsandkeys/ballerinaTruststore.p12"
 icpServicePort = 9165
 
-[ballerinax.wso2.controlplane.dashboard]
+[ballerinax.wso2.icp.dashboard]
 url = "https://localhost:9743/dashboard/api"
 heartbeatInterval = 5
 groupId = "bal_dev"
